requirements.txt
asgiref==3.2.3
autopep8==1.4.4
Django==3.0.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
drf-base64==2.0
freeze==1.0.10
Pillow==7.0.0
pkg-resources==0.0.0
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
